Add status to facility list page
The facility list details page shows the status of the list. Users
are able to see additional information in a dialog if a list is
rejected. Admin users are given approve and reject buttons, and the
ability to add a status change reason for rejections.

Note that the actions for the approve/reject buttons are stubs, and the
actions will be integrated in issue #2018.