Handle facility loading appropriately on mount
When loading the homepage, we don't fetch the facilities list
because it isn't displayed on the homepage.

If there is an existing querystring when the homepage is loaded, we
replace the route with the facilities route. In that case, the
pathname is sometimes appended with a forward slash, which was
causing the test for if we should fetch the list to fail. Cleaning
the pathname resolves this.