Rewrite the GazetteerCache to work with database-backed gazetteer
Our new gazetteer subclasses use the database for sharing trained models and
index data so the code within the GazetteerCache used to keep the in-memory
model in sync with database changes is no longer needed.

Instead, the GazetteerCache now provides class methods that wrap the gazetteer
instance methods and handle refreshing the gazetteer only if a new one has been
trained and activated.