Provision an ElastiCache memcached cluster
We are using memcached as a cache backend to enable API request throttling.

This uses a pre-release version of terraform-aws-memcached-elasticache module
with suupport for Terraform version 0.12+

The values for ec_memcached_max_item_size, alarm_cpu_threshold_percent, and
alarm_memory_threshold_bytes were copied from the default values. The
ec_memcached_maintenance_window was chosen based on a previous project value.
ElastiCache requires a minimum 1-hour window.

We chose the lastest version of memcached available at the time of writing.

The cache.t3.medium size was chosen because it offers an ample 3GB of memory and
2 vCPUs at $1.63 a day. We choose 1 cluster node to keep the cluster logic
simple since it is not immediately clear if the old django-elasticace library
recommended by the AWS docs will still work correctly.