Delete additional matches on facility delete
If a facility has more than one match, those other matches could be
promoted when the original facility is deleted. This means that a
contributer might need to delete a facility many times to fully delete
the facility.

In order to simplify the delete process, facility list items from the
same contributor as the original facility should be deleted along with
the original facility. This includes the matches associated with these
facility list items.

This commit accomplishes this by adding facility list items from the
same contributor as the created_from contributor of the facility to be
deleted to the list of facility list items to delete.

Additionally, when looking for a best match to promote, facilitiy list
items from the original contributor are excluded. If no best match is
found, nothing changes since all matches and list items are deleted
anyway. If a best match is found, any matches from the original
contributor are deleted along with their list items.