Update sector/product type parsing
Sector/product type values should be compared against those in the
database. If there is a match, a sector is added. Otherwise a product
type is added.

Some tests needed to be updated for these changes. For example, Beauty
is not in the most recent sectors.csv list, so it shouldn't be used as
a sector test value. Also, the addition of a default sector value,
Unspecified, means sector is no longer required in the create API
request body.