Add new subscribers to HubSpot
Updates the `add_user_to_mailing_list` function to add new users to
HubSpot instead of MailChimp. Unlike with MailChimp, this is a
multi-step process. First, the user is added as a Contact. The
HubSpot Private App is set in the account configuration to automatically
set all new contacts who are added via the private app (our API key)
as marketing contacts. Next, the new contact is subscribed to the
subscription type indicated by the API key. Dynamic lists which are set
to filter contacts by that subscription type then automatically add the
subscribed user to the appropriate list(s).