Add sector/product type parse logic to csv import
This commit extracts the sector/product type parsing to a separate file
with a base class and two implementers of that class for Requests and
CSV imports.

The reason for moving this functionality out of views.py is two fold.
First, the code is repeated in both views.py and in processing.py for
csv imports. Also, csv imports might require processing many rows. Doing
a Sector look up for each row is inefficient and could add a lot of
processing time.

This commit deletes test_sector_invalid_value from
FacilityListItemParseTest. This is because sectors can no longer have a
value with more than 50 characters since sector can only be one of the
values found in the Sector DB table.