Update footer for OS Hub
These changes are mostly contained to the footer except for a few parts:

The <main> component's position has been changed from 'fixed' to
'default'.

The header's zIndex has been increased to appear over the map which can
now scroll under the header.

The FilterSidebar and results on the facilities page have been given a
max height to keep them from stretching the map to their length.