Add useHeightSubtract
This commit extracts the functionality used to determine the result list
height into a hook. It's generalized and other components use this to
determine their heights.

This is not ideal, but the InfiniteAnyHeight component requires an
absolute height, so the rest of the components in the main grid require
absolute heights as well.