Handle empty facility_types
When a facility type or processing type is submitted with a pipe as
the final character, it creates an empty string as the final
value in the array. When we attempted to process the empty string,
it was cleaned to a None value, which threw an error. Empty strings are
now processed as 'No match' or (None, None, None, None).