Only use facility/processing taxonomy for apparel
Facilities with 'Apparel' as one of their sectors are considered to
be in the Apparel sector.

The taxonomy was created for the apparel sector.
Facilities outside of that sector shouldn't be assigned a
facility_type, and any submitted facility_type_processing_type
should be applied as a processing_type without matching to the
taxonomy.