Add management command to replace oar_id in JSON fields
The `FacilityListItem` and `FacilityMatch` models both have JSON fields which
can contain keys that reference oar_id rather than os_id. Splitting a facility
was failing with an error due to the code expecting a key name containing os_id.

Despite the fact that it is a one time operation, we wrote this as a management
command rather than a data migration to facilitate testing with a "dry run"
option.