Fix map not appearing at medium widths
When we introduced the new footer we included a `useMaoHeight` hook that would
provide a calculated value for the `Map` component. The `Map` component created
in `MapAndSidebar` made use of this hook but the `Map` component created in
`FilterSidebar` did not, resulting in the map not appearing at medium scree
widths.