Fix removed list item filtering
There are now two types of removed items: ones that are removed after
processing (which are indicated by having no active matches), and ones
that are removed prior to processing (which are indicated by the
'ITEM_REMOVED' status). Adjust the filtering for the list details page
to include both types when filtering by 'REMOVED'.