Prevent infinite loading on error
The sidebar error boundary previously responded to errors by
attempting to reload the components which had thrown the error.
This resulted in the page infinitely loading.

There are now two levels of error boundaries for the sidebar(s). One
handles errors in the filters column and search results column; the
other handles errors in the combined sidebar component.

Both levels of error boundary will show an error message instead of the
effected component when an error has occurred.