Remove most production resources
we're updating the terraform to remove all resources from the OAR
production environment, except for cloudfront and its related
DNS and certificate resources.

The infra stage of the Jenkinsfile has been removed. Because the
staging environment no longer exists, if this hadn't been done,
merging a PR into the development branch would recreate staging
resources, which is not what we want.